floorplan_parser/requirements.txt
Pillow>=10.0.0
google-generativeai>=0.7.0
pydantic>=2.0.0
opencv-python>=4.8.0
numpy>=1.26.0
